requirements.txt
python-telegram-bot==21.3
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
